// pong/src/utils/Types.ts
export interface DepthLineBehavior {
    movement?: 'vertical' | 'horizontal' | string;
    direction?: 'upwards' | 'downwards' | 'left' | 'right' | string;
    fade?: 'in' | 'out' | 'none' | string;
    
	// Pyramid specific properties
    pyramidBaseHeight?: number;
    pyramidBaseWidth?: number;
    pyramidPeakHeight?: number;
    pyramidPeakOffset?: number;

	// Ruin specicif properties
	ruinHSegments?: number;
	ruinTSegments?: number;
	maxHeight?: number;
	maxWidth?: number;
	segmentWidths?: number[];
    segmentHeights?: number[];
	hOffset?: number;

	// Wave specific properties
    waveAmplitude?: number;
    waveFrequency?: number;
    waveOffset?: number;
    phaseShift?: number;
    waveType?: 'sine' | 'square' | 'triangle' | 'sawtooth';
    amplitudeVariation?: number;   // Add this property
    minBottomHeight?: number | null;  // Add this property
}

export type FrameData = {
	deltaTime: number;
};
// ...
export type GameEvent = {
    type: string;
    side?: 'left' | 'right'; // For example, for the SCORE event
	target?: Paddle | World;
	entitiesMap?: Map<string, Entity>;
	points?: Point[];
	x?: number;
	y?: number;
    // Add any other fields you expect in an event here
};
// ...
export interface World {
	theme: string;
	name: string;
	color: number;
}